pyproject.toml
[project]
name = "langgraph-research"
version = "0.1.0"
description = "A research tool for automated literature review generation"
authors = [
    {name = "franciszek-latala",email = "Franciszek.Latala@crowe-peak.nl"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "langgraph (>=0.5.2)",
    "langchain-openai (>=0.3.27)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-core (>=0.3.68)",
    "langchain-community (>=0.3.27)",
    "langchain-tavily (>=0.2.0)",
    "tavily-python (>=0.5.0)",
    "arxiv (>=2.2.0,<3.0.0)",
    "pymupdf (>=1.26.3,<2.0.0)",
    "sqlalchemy (>=2.0.0,<3.0.0)",
    "chromadb (>=0.5.0,<1.0.0)",
    "rich (>=14.2.0,<15.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false